requirements.txt
llama-index==0.12.5
llama-index-vector-stores-qdrant==0.4.3
llama-index-embeddings-ollama==0.5.0
llama-index-llms-ollama==0.5.0
qdrant-client==1.17.0
python-dotenv==1.1.0
pyyaml==6.0.2
beautifulsoup4==4.13.4
pytest==8.3.5
